+++ packages/cli/src/BootCommand.ts
@@ -6,7 +6,8 @@ import React from "react";
 import { render } from "ink";
 import { Config, DbProviderType, GraphProviderType } from "@bb/types";
 import { HINTS, getBytebellHome, getConfigValue, isDevMode } from "@bb/config";
-import { applyInfraDefaults, checkPreflight } from "./bootConfig.ts";
+import { checkPreflight } from "./bootConfig.ts";
+import { applyInfraDefaults } from "./infraDefaults.ts";
 import { SetupForm } from "./SetupForm.tsx";
 import { error, info, success } from "./output.ts";
 import { bringInfraUp, usingHonker } from "./bootInfra.ts";
